agents/social/requirements.txt
# /Users/linchr/Desktop/work/datecoach/agents/social/requirements.txt
google-cloud-spanner
google-generativeai
google-adk
python-dotenv
fastapi 
uvicorn
a2a_common-0.1.0-py3-none-any.whl
